replace dataSource / add bottom menu / rename popylarMovie -> releasedMovie
