+++ app/src/main/java/com/example/filmoteca/data/datasources/remote/movie/MovieRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.example.filmoteca.data.datasources.remote.movie
+
+import com.example.filmoteca.data.remote.dto.MovieResponseDto
+
+interface MovieRemoteDataSource {
+    suspend fun getReleasedMovies(): MovieResponseDto
+}
