--- app/src/main/java/com/example/filmoteca/data/remote/datasource/MovieRemoteDataSource.kt
@@ -1,10 +1,0 @@
-// data/remote/MovieRemoteDataSource.kt
-package com.example.filmoteca.data.remote
-
-import com.example.filmoteca.data.remote.api.MovieApiService
-import com.example.filmoteca.data.remote.dto.MovieResponseDto
-
-interface MovieRemoteDataSource {
-    suspend fun getPopularMovies(): MovieResponseDto
-}
-
