+++ app/src/main/java/com/example/filmoteca/data/remote/dto/MovieResponseDto.kt
@@ -3,9 +3,9 @@ package com.example.filmoteca.data.remote.dto
 import com.google.gson.annotations.SerializedName
 
 class MovieResponseDto (
-    val page: Int = 0,
+    val page: Int,
     val results: List<MovieDto>,
-    @SerializedName("total_pages") val totalPages: Int = 0,
-    @SerializedName("total_results") val totalResults: Int = 0,
+    @SerializedName("total_pages") val totalPages: Int,
+    @SerializedName("total_results") val totalResults: Int,
 
 )
