+++ app/src/main/java/com/example/filmoteca/domain/movie/repository/MovieRepository.kt
@@ -5,5 +5,5 @@ import com.example.filmoteca.domain.movie.model.Movie
 import kotlinx.coroutines.flow.Flow
 
 interface MovieRepository {
-    fun getPopularMovies(): Flow<List<Movie>>
+    fun getReleasedMovies(): Flow<List<Movie>>
 }
