+++ app/src/main/java/com/example/filmoteca/domain/movie/usecase/GetReleasedMoviesUseCase.kt
@@ -4,7 +4,7 @@ import com.example.filmoteca.domain.movie.model.Movie
 import com.example.filmoteca.domain.movie.repository.MovieRepository
 import kotlinx.coroutines.flow.Flow
 
-class GetPopularMoviesUseCase(private val repository: MovieRepository) {
-    operator fun invoke(): Flow<List<Movie>> = repository.getPopularMovies()
+class GetReleasedMoviesUseCase(private val repository: MovieRepository) {
+    operator fun invoke(): Flow<List<Movie>> = repository.getReleasedMovies()
 }
