+++ app/src/main/java/com/example/filmoteca/ui/MainActivity.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.GridLayoutManager
-import com.example.filmoteca.data.remote.datasourceimpl.MovieRemoteDataSourceImpl
+import com.example.filmoteca.data.datasources.remote.movie.MovieRemoteDataSourceImpl
 import com.example.filmoteca.data.remote.network.RetrofitInit
 import com.example.filmoteca.data.repositoryimpl.MovieRepositoryImpl
 import com.example.filmoteca.databinding.ActivityMainBinding
@@ -19,39 +19,6 @@ import com.example.filmoteca.presentation.viewmodel.MovieListViewModel
 import com.example.filmoteca.presentation.viewmodel.MovieListViewModelFactory
 import com.example.filmoteca.ui.movie.adapter.MovieAdapter
 import kotlinx.coroutines.launch
-
-//class MainActivity : AppCompatActivity() {
-//    private lateinit var binding: ActivityMainBinding
-//    private lateinit var adapter: MovieAdapter
-//    private lateinit var viewModel: MovieListViewModel
-//
-//    override fun onCreate(savedInstanceState: Bundle?) {
-//        super.onCreate(savedInstanceState)
-//        enableEdgeToEdge()
-//        binding = ActivityMainBinding.inflate(layoutInflater)
-//        setContentView(binding.root)
-//
-//
-//
-//        val api = RetrofitInit.api
-//        val repository = MovieRepositoryImpl(api)
-//        val factory = MovieListViewModelFactory(repository)
-//        viewModel = ViewModelProvider(this, factory)[MovieListViewModel::class.java]
-//
-//
-//        binding.moviesRecyclerView.layoutManager = GridLayoutManager(this, 2)
-//        adapter = MovieAdapter(emptyList())
-//        binding.moviesRecyclerView.adapter = adapter
-//
-//        viewModel.movies.observe(this) { movies ->
-//            adapter.setItems(movies)
-//        }
-//        viewModel.loadMovies()
-//    }
-//}
-
-
-
 
 class MainActivity : AppCompatActivity() {
 
@@ -96,13 +63,6 @@ class MainActivity : AppCompatActivity() {
 
         // Загрузка данных
         viewModel.loadMovies()
-
-
-
     }
-
-
-
-
 
 }
