+++ app/src/main/res/menu/bottom_nav.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+
+    <item
+        android:id="@+id/"
+        android:icon=""
+        />
+
+    <item
+        android:id="@+id/mainScreen"
+        android:icon=""
+        />
+
+    <item
+        android:id="@+id/profile"
+        android:icon= ""
+        />
+</menu>
